Add notifications view to display medication expiration alerts
